add int price to recipes

+++ database/migrations/2018_08_13_155654_add_price_to_recipes.php
@@ -15,7 +15,7 @@ class AddPriceToRecipes extends Migration
     {
         Schema::table('recipes', function (Blueprint $table) {
 
-            $table->float('price')->after('Level')->default(0);
+            $table->integer('price')->after('Level')->default(0);
         });
     }
